evals/pyproject.toml
[project]
name = "agentregistry-evals"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "google-adk[a2a]",
    "pytest",
    "pytest-asyncio",
    "httpx",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
